requirements.txt
setuptools~=69.1.1
tqdm~=4.66.2
PyPDF2~=3.0.1
moviepy~=1.0.3
pydub~=0.25.1
schedule~=1.2.1
pytube~=15.0.0
pytube3
youtube_dl
pdf2docx~=0.5.8
simple-image-download
colorama~=0.4.6
openpyxl~=3.1.5
psutil~=6.0.0
pillow~=10.3.0
fpdf~=1.7.2
Pyment~=0.3.3